Move gitlab-related stuff to a separate module
